requirements.txt
streamlit==1.48.0
pandas==2.0.0
openpyxl==3.1.0
numpy==1.24.0
sqlalchemy==2.0.0
python-dateutil==2.8.0
bcrypt==3.1.7
pyyaml==6.0
altair==5.0.0
streamlit-authenticator==0.2.3
